requirements.txt
annotated-types==0.7.0
anyio==4.9.0
backoff==2.2.1
certifi==2025.6.15
charset-normalizer==3.4.2
distro==1.9.0
dotenv==0.9.9
gitdb==4.0.12
GitPython==3.1.41
greenlet==3.2.3
grpcio==1.73.0
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
jiter==0.10.0
mem0ai==0.1.108
numpy==2.3.0
openai==1.88.0
portalocker==2.10.1
posthog==5.0.0
protobuf==6.31.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
qdrant-client==1.14.3
requests==2.32.4
setuptools==75.6.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
